Create distance detector
Just a somewhat improvement from the basic motion sensor of the other file in the repository. Only difference is that an LCD screen and a 3 LED's were attached instead with the distance and the color displayed don the LCD and the LED's flashing in correspondence to the distance the motion sensor detects.